+++ setup.py
@@ -143,6 +143,7 @@ def configuration(parent_package='', top_path=None):
 
     config.add_subpackage('wradlib')
     config.add_subpackage('wradlib.io')
+    config.add_subpackage('wradlib.georef')
 
     return config
